+++ crate/server/src/database/tests/find_attributes_test.rs
@@ -18,7 +18,11 @@ use cosmian_kms_utils::{
 use cosmian_logger::log_utils::log_init;
 use uuid::Uuid;
 
-use crate::{database::Database, kms_bail, result::KResult};
+use crate::{
+    database::{object_with_metadata::ObjectWithMetadata, Database},
+    kms_bail,
+    result::KResult,
+};
 
 pub async fn find_attributes<DB: Database>(
     db_and_params: &(DB, Option<ExtraDatabaseParams>),
@@ -60,7 +64,10 @@ pub async fn find_attributes<DB: Database>(
 
     let objs_ = db
         .retrieve(&uid, owner, ObjectOperationType::Get, db_params)
-        .await?;
+        .await?
+        .into_values()
+        .collect::<Vec<ObjectWithMetadata>>();
+
     assert_eq!(objs_.len(), 1);
     match objs_.len() {
         1 => {